+++ src/detectors/pca_model.cpp
@@ -74,9 +74,9 @@ void PcaModel::Load(const std::string& artifact_path) {
     // 2. PCA
     // Components are stored as list of lists (k x d)
     auto raw_components = j["model"]["components"].get<std::vector<std::vector<double>>>();
-    int k = raw_components.size();
+    int k = static_cast<int>(raw_components.size());
     if (k == 0) throw std::runtime_error("No PCA components found");
-    int d = raw_components[0].size();
+    int d = static_cast<int>(raw_components[0].size());
     if (d != FeatureVector::kSize) throw std::runtime_error("Dimension mismatch in PCA components");
 
     // Copy to matrix
